add gen_template_docs.py to generate docs
Import Kyle Liberti's gen_template_docs.py which can generate template
documentation from the template definitions. Some minor adjustments to
run the script from this repository:

 * swap out tab characters and use 4-spaces for indent
 * remove code to clone this repo, since the script is part of it
 * work on the assumption we're running from the repo (special
   case: if $PWD is 'docs', change down a directory)
 * Add a check to catch missing labels in the JSON for the secrets
   templates
 * Add auto-generation of a template documentation index page
 * introduce template.adoc.in and index.adoc.in to remove hard-coded
   asciidoc page templates from the script

For now, the ./secrets directory is ignored, since we currently do not
generate documentation output for it due to missing labels. This
prevents us generating a link to the non-existent documentation on the
index page.