docs: rework using python ptemplate library
Rework the python doc generation script using the ptemplate library,
rather than the library-supplied string.Template. This allows us to
optionally enable sections for different image types. To install it
locally:

    pip install --user ptemplate

Only generate a section if there is data in the JSON file for it.

Add some metadata to the secrets templates, and remove them from the
generator script's ignore list.

Add dward's clustering paragraph to the template page. Put it in its
own section and only output for EAP templates.

Do not output the Secrets section for the secrets templates.

Adjust the readinessProbe section to output preformatted text rather
than a table.

Fix the table-of-contents for generated pages.