gen_template_docs: adjust input template selection
Adjust how gen_template_docs finds templates. Instead of maintaining
a blacklist of directories *not* to check, maintain a list of
directories known to contain templates (eap, amq, secrets, webserver).

Allow the user to specify which templates to generate documentation
for. If any command-line arguments are supplied, interpret them as
the templates to read (and don't go looking for others).
Useful for debugging when a particular input has broken the generator.