HTTPS_SECRET is required as the field can't be blank
Currently, HTTPS_SECRET parameter is not required but it has a default
value. If the default value is removed and the field is blank, the app
deployment will fail as it needs the value of the field. As a result,
the parameter should be marked as required to avoid users to remove
the default value.

The parameter is changed to required: "true" to enforce users to either
keep the default value or provide a new value and can't keep the field
blank to avoid failure.

Signed-off-by: Vu Dinh <vdinh@redhat.com>